requirements.txt
# requirements.txt
snowflake-snowpark-python
streamlit